07_tracer/pyproject.toml
[project]
name = "tracer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
infra = [
    "aws-cdk-lib==2.215.0",
    "constructs>=10.0.0,<11.0.0",
]
lambda = [
    "aws-lambda-powertools>=3.22.0",
]
